build: reorganize files [SF-246]
