+++ packages/sf-graph-client/jest.config.ts
@@ -8,4 +8,8 @@ module.exports = {
     collectCoverageFrom: ['src/**/*.ts', '!**/*.module.ts', '!**/index.ts'],
     coverageDirectory: './coverage',
     testEnvironment: 'node',
+    rootDir: '.',
+    moduleNameMapper: {
+        'src/(.*)': '<rootDir>/src/$1',
+    },
 };
