+++ packages/sf-graph-client/src/hooks/index.ts
@@ -1,1 +1,4 @@
-export * from './lending-market';
+export * from './types';
+export * from './useLendingMarkets';
+export * from './useQuery';
+export * from './useTransactionHistory';
