--- packages/sf-graph-client/src/hooks/lending-market/index.ts
@@ -1,3 +1,0 @@
-export * from './useLendingMarkets';
-export * from './useOrderHistory';
-export * from './useTransactionHistory';
