+++ packages/sf-graph-client/src/hooks/useTransactionHistory/index.ts
@@ -1,0 +1,1 @@
+export * from './useTransactionHistory';
